requirements.txt
anyio==4.4.0
click==8.1.7
graphql-core==3.2.4
h11==0.14.0
idna==3.8
libcst==1.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
strawberry-graphql==0.239.2
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.30.6
